@@ -1,6 +1,4 @@
-import 'package:flutter/foundation.dart';
-
-class Product{
+class Product {
   final String productName;
   final String unitType;
   final double unitPrice;
@@ -12,39 +10,43 @@ class Product{
   final String note;
 
   Product({
-    @required this.approved,
-    @required this.availableUnits,
-    this.imageUrl = "",
-    this.note = "",
-    @required this.productId,
-    @required this.productName,
-    @required this.unitPrice, 
-    @required this.unitType,
-    @required this.vendorId
+    required this.approved,
+    required this.availableUnits,
+    this.imageUrl = '',
+    this.note = '',
+    required this.productId,
+    required this.productName,
+    required this.unitPrice,
+    required this.unitType,
+    required this.vendorId,
   });
 
   Map<String, dynamic> toMap() {
     return {
-      'productName' : productName,
-      'unitType' : unitType,
-      'unitPrice' : unitPrice,
+      'productName': productName,
+      'unitType': unitType,
+      'unitPrice': unitPrice,
       'availableUnits': availableUnits,
       'approved': approved,
-      'imageUrl':imageUrl,
-      'note':note,
-      'productId':productId,
-      'vendorId':vendorId
+      'imageUrl': imageUrl,
+      'note': note,
+      'productId': productId,
+      'vendorId': vendorId,
     };
   }
 
-  Product.fromFirestore(Map<String, dynamic> firestore)
-    : productName = firestore['productName'],
-      unitType = firestore['unitType'],
-      unitPrice = firestore['unitPrice'],
-      availableUnits = firestore['availableUnits'],
-      approved = firestore['approved'],
-      imageUrl = firestore['imageUrl'],
-      note = firestore['note'],
-      productId = firestore['productId'],
-      vendorId = firestore['vendorId'];
+  factory Product.fromFirestore(Map<String, dynamic> firestore) {
+    final unitPriceRaw = firestore['unitPrice'];
+    return Product(
+      productName: (firestore['productName'] ?? '').toString(),
+      unitType: (firestore['unitType'] ?? '').toString(),
+      unitPrice: unitPriceRaw is num ? unitPriceRaw.toDouble() : 0.0,
+      availableUnits: (firestore['availableUnits'] ?? 0) as int,
+      approved: (firestore['approved'] ?? false) as bool,
+      imageUrl: (firestore['imageUrl'] ?? '').toString(),
+      note: (firestore['note'] ?? '').toString(),
+      productId: (firestore['productId'] ?? '').toString(),
+      vendorId: (firestore['vendorId'] ?? '').toString(),
+    );
+  }
 }
