@@ -1,4 +1,5 @@
 import 'dart:async';
+import 'dart:io';
 
 import 'package:farmers_market/src/blocs/auth_bloc.dart';
 import 'package:farmers_market/src/styles/tabbar.dart';
@@ -9,22 +10,20 @@ import 'package:farmers_market/src/widgets/profile.dart';
 import 'package:farmers_market/src/widgets/vendor_scaffold.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'dart:io';
-
 import 'package:provider/provider.dart';
 
 class Vendor extends StatefulWidget {
-  
+  const Vendor({super.key});
 
   @override
-  _VendorState createState() => _VendorState();
+  State<Vendor> createState() => _VendorState();
 
   static TabBar get vendorTabBar {
     return TabBar(
-      unselectedLabelColor: TabBarStyles.unselectedLabelColor ,
-      labelColor: TabBarStyles.labelColor ,
-      indicatorColor: TabBarStyles.indicatorColor ,
-      tabs: <Widget>[
+      unselectedLabelColor: TabBarStyles.unselectedLabelColor,
+      labelColor: TabBarStyles.labelColor,
+      indicatorColor: TabBarStyles.indicatorColor,
+      tabs: const <Widget>[
         Tab(icon: Icon(Icons.list)),
         Tab(icon: Icon(Icons.shopping_cart)),
         Tab(icon: Icon(Icons.person)),
@@ -34,58 +33,64 @@ class Vendor extends StatefulWidget {
 }
 
 class _VendorState extends State<Vendor> {
-  StreamSubscription _userSubscription;
+  StreamSubscription? _userSubscription;
 
   @override
   void initState() {
-    Future.delayed(Duration.zero, (){ 
-        var authBloc = Provider.of<AuthBloc>(context,listen: false);
-        _userSubscription = authBloc.user.listen((user) { 
-          if (user == null) Navigator.of(context).pushNamedAndRemoveUntil('/login', (route) => false);
-        });
-    });
-   
     super.initState();
+    Future.delayed(Duration.zero, () {
+      final authBloc = Provider.of<AuthBloc>(context, listen: false);
+      _userSubscription = authBloc.user.listen((user) {
+        if (user == null) {
+          Navigator.of(context)
+              .pushNamedAndRemoveUntil('/login', (route) => false);
+        }
+      });
+    });
   }
 
   @override
   void dispose() {
-    _userSubscription.cancel();
+    _userSubscription?.cancel();
     super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
-    
     if (Platform.isIOS) {
-      return CupertinoPageScaffold(  
+      return CupertinoPageScaffold(
         child: NestedScrollView(
-          headerSliverBuilder: (BuildContext context, bool innerBoxIsScrolled){
-            return <Widget> [
-              AppNavbar.cupertinoNavBar(title: 'Vendor Name',context: context),
+          headerSliverBuilder: (BuildContext context, bool innerBoxIsScrolled) {
+            return <Widget>[
+              AppNavbar.cupertinoNavBar(title: 'Vendor Name', context: context),
             ];
-          }, 
+          },
           body: VendorScaffold.cupertinoTabScaffold,
-      ),
-      );
-    } else {
-      return DefaultTabController(  
-        length: 3,
-        child: Scaffold(  
-          body: NestedScrollView(  
-            headerSliverBuilder: (BuildContext context, bool innerBoxIsScrolled){
-              return <Widget> [
-                AppNavbar.materialNavBar(title: 'Vendor Name', tabBar: Vendor.vendorTabBar)
-              ];
-            },
-            body: TabBarView(children: <Widget>[
-              Products(),
-              Orders(),
-              Profile(),
-            ],)
-          )
         ),
       );
     }
+
+    return DefaultTabController(
+      length: 3,
+      child: Scaffold(
+        body: NestedScrollView(
+          headerSliverBuilder: (BuildContext context, bool innerBoxIsScrolled) {
+            return <Widget>[
+              AppNavbar.materialNavBar(
+                title: 'Vendor Name',
+                tabBar: Vendor.vendorTabBar,
+              )
+            ];
+          },
+          body: TabBarView(
+            children: <Widget>[
+              Products(),
+              Orders(),
+              const Profile(),
+            ],
+          ),
+        ),
+      ),
+    );
   }
 }
