@@ -4,93 +4,93 @@ import 'package:farmers_market/src/styles/colors.dart';
 import 'package:farmers_market/src/styles/text.dart';
 import 'package:flutter/material.dart';
 
-class AppButton extends StatefulWidget{
+class AppButton extends StatefulWidget {
   final String buttonText;
   final ButtonType buttonType;
-  final void Function() onPressed;
+  final VoidCallback? onPressed;
 
-  AppButton({
-    @required this.buttonText,
-    this.buttonType,
-    this.onPressed
+  const AppButton({
+    super.key,
+    required this.buttonText,
+    this.buttonType = ButtonType.LightBlue,
+    this.onPressed,
   });
 
   @override
-  _AppButtonState createState() => _AppButtonState();
+  State<AppButton> createState() => _AppButtonState();
 }
 
 class _AppButtonState extends State<AppButton> {
   bool pressed = false;
 
+  Color _buttonColor(ButtonType type) {
+    switch (type) {
+      case ButtonType.Straw:
+        return AppColors.straw;
+      case ButtonType.LightBlue:
+        return AppColors.lightblue;
+      case ButtonType.DarkBlue:
+        return AppColors.darkblue;
+      case ButtonType.Disabled:
+        return AppColors.lightgray;
+      case ButtonType.DarkGray:
+        return AppColors.darkgray;
+    }
+  }
+
   @override
   Widget build(BuildContext context) {
-    TextStyle fontStyle;
-    Color buttonColor;
-
-    switch (widget.buttonType) {
-      case ButtonType.Straw:
-      fontStyle = TextStyles.buttonTextLight;
-      buttonColor = AppColors.straw;
-      break;
-       case ButtonType.LightBlue:
-      fontStyle = TextStyles.buttonTextLight;
-      buttonColor = AppColors.lightblue;
-      break;
-       case ButtonType.DarkBlue:
-      fontStyle = TextStyles.buttonTextLight;
-      buttonColor = AppColors.darkblue;
-      break;
-       case ButtonType.Disabled:
-      fontStyle = TextStyles.buttonTextLight;
-      buttonColor = AppColors.lightgray;
-      break;
-       case ButtonType.DarkGray:
-      fontStyle = TextStyles.buttonTextLight;
-      buttonColor = AppColors.darkgray;
-      break;
-      default: 
-      fontStyle = TextStyles.buttonTextLight;
-      buttonColor = AppColors.lightblue;
-      break;
-    }
+    final buttonColor = _buttonColor(widget.buttonType);
 
     return AnimatedContainer(
-      padding: EdgeInsets.only( 
-        top: (pressed) ? BaseStyles.listFieldVertical + BaseStyles.animationOffset :  BaseStyles.listFieldVertical,
-        bottom: (pressed) ? BaseStyles.listFieldVertical - BaseStyles.animationOffset :  BaseStyles.listFieldVertical,
+      padding: EdgeInsets.only(
+        top: pressed
+            ? BaseStyles.listFieldVertical + BaseStyles.animationOffset
+            : BaseStyles.listFieldVertical,
+        bottom: pressed
+            ? BaseStyles.listFieldVertical - BaseStyles.animationOffset
+            : BaseStyles.listFieldVertical,
         left: BaseStyles.listFieldHorizontal,
-        right: BaseStyles.listFieldHorizontal
+        right: BaseStyles.listFieldHorizontal,
       ),
+      duration: const Duration(milliseconds: 20),
       child: GestureDetector(
-              child: Container(  
-          height: ButtonStyles.buttonHeight,
-          width: MediaQuery.of(context).size.width,
-          decoration: BoxDecoration(  
-            color: buttonColor,
-            borderRadius: BorderRadius.circular(BaseStyles.borderRadius),
-            boxShadow: pressed ? BaseStyles.boxShadowPressed : BaseStyles.boxShadow
-          ),
-          child: Center(child: Text(widget.buttonText,style: fontStyle,)),
-        ),
-        onTapDown: (details){
-          setState(() {
-            if (widget.buttonType != ButtonType.Disabled) pressed = !pressed;
-          });
-        },
-        onTapUp: (details){
-          setState(() {
-            if (widget.buttonType != ButtonType.Disabled) pressed = !pressed;
-          });
-        },
-        onTap: (){
+        onTapDown: (_) {
           if (widget.buttonType != ButtonType.Disabled) {
-            widget.onPressed();
+            setState(() => pressed = true);
           }
         },
+        onTapCancel: () {
+          if (widget.buttonType != ButtonType.Disabled) {
+            setState(() => pressed = false);
+          }
+        },
+        onTapUp: (_) {
+          if (widget.buttonType != ButtonType.Disabled) {
+            setState(() => pressed = false);
+          }
+        },
+        onTap: () {
+          if (widget.buttonType != ButtonType.Disabled) {
+            widget.onPressed?.call();
+          }
+        },
+        child: Container(
+          height: ButtonStyles.buttonHeight,
+          width: MediaQuery.of(context).size.width,
+          decoration: BoxDecoration(
+            color: buttonColor,
+            borderRadius: BorderRadius.circular(BaseStyles.borderRadius),
+            boxShadow:
+                pressed ? BaseStyles.boxShadowPressed : BaseStyles.boxShadow,
+          ),
+          child: Center(
+            child: Text(widget.buttonText, style: TextStyles.buttonTextLight),
+          ),
+        ),
       ),
-      duration: Duration(milliseconds: 20),
     );
   }
 }
 
-enum ButtonType { LightBlue,Straw, Disabled, DarkGray, DarkBlue }
+enum ButtonType { LightBlue, Straw, Disabled, DarkGray, DarkBlue }
