@@ -1,26 +1,29 @@
+import 'dart:io';
+
 import 'package:farmers_market/src/styles/base.dart';
 import 'package:farmers_market/src/styles/buttons.dart';
 import 'package:farmers_market/src/styles/colors.dart';
 import 'package:farmers_market/src/styles/text.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'dart:io';
 
 class AppDropdownButton extends StatelessWidget {
   final List<String> items;
   final String hintText;
   final IconData materialIcon;
   final IconData cupertinoIcon;
-  final String value;
-  final Function(String) onChanged;
+  final String? value;
+  final ValueChanged<String?>? onChanged;
 
-  AppDropdownButton(
-      {@required this.items,
-      @required this.hintText,
-      this.materialIcon,
-      this.cupertinoIcon,
-      this.value,
-      this.onChanged});
+  const AppDropdownButton({
+    super.key,
+    required this.items,
+    required this.hintText,
+    required this.materialIcon,
+    required this.cupertinoIcon,
+    this.value,
+    this.onChanged,
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -31,54 +34,30 @@ class AppDropdownButton extends StatelessWidget {
           height: ButtonStyles.buttonHeight,
           width: MediaQuery.of(context).size.width,
           decoration: BoxDecoration(
-              borderRadius: BorderRadius.circular(BaseStyles.borderRadius),
-              border: Border.all(
-                  color: AppColors.straw, width: BaseStyles.borderWidth)),
-          child: Row(
-            children: <Widget>[
-              Container(
-                  width: 35.0, child: BaseStyles.iconPrefix(materialIcon)),
-              Expanded(
-                child: Center(
-                    child: GestureDetector(
-                  child:(value == null) ? Text(hintText, style: TextStyles.suggestion)
-                  : Text(value, style: TextStyles.body),
-                  onTap: () {
-                    showCupertinoModalPopup(
-                        context: context, builder: (BuildContext context) {
-                          return _selectIOS(context, items, value);
-                        });
-                  },
-                )),
-              ),
-            ],
+            borderRadius: BorderRadius.circular(BaseStyles.borderRadius),
+            border: Border.all(
+              color: AppColors.straw,
+              width: BaseStyles.borderWidth,
+            ),
           ),
-        ),
-      );
-    } else {
-      return Padding(
-        padding: BaseStyles.listPadding,
-        child: Container(
-          height: ButtonStyles.buttonHeight,
-          width: MediaQuery.of(context).size.width,
-          decoration: BoxDecoration(
-              borderRadius: BorderRadius.circular(BaseStyles.borderRadius),
-              border: Border.all(
-                  color: AppColors.straw, width: BaseStyles.borderWidth)),
           child: Row(
             children: <Widget>[
-              Container(
-                  width: 35.0, child: BaseStyles.iconPrefix(materialIcon)),
+              SizedBox(width: 35.0, child: BaseStyles.iconPrefix(materialIcon)),
               Expanded(
                 child: Center(
-                  child: DropdownButton<String>(
-                    items: buildMaterialItems(items),
-                    value: value,
-                    hint: Text(hintText, style: TextStyles.suggestion),
-                    style: TextStyles.body,
-                    underline: Container(),
-                    iconEnabledColor: AppColors.straw,
-                    onChanged: (value)  => onChanged(value),
+                  child: GestureDetector(
+                    child: Text(
+                      value ?? hintText,
+                      style: (value == null)
+                          ? TextStyles.suggestion
+                          : TextStyles.body,
+                    ),
+                    onTap: () {
+                      showCupertinoModalPopup(
+                        context: context,
+                        builder: (_) => _selectIOS(context),
+                      );
+                    },
                   ),
                 ),
               ),
@@ -87,18 +66,46 @@ class AppDropdownButton extends StatelessWidget {
         ),
       );
     }
+
+    return Padding(
+      padding: BaseStyles.listPadding,
+      child: Container(
+        height: ButtonStyles.buttonHeight,
+        width: MediaQuery.of(context).size.width,
+        decoration: BoxDecoration(
+          borderRadius: BorderRadius.circular(BaseStyles.borderRadius),
+          border:
+              Border.all(color: AppColors.straw, width: BaseStyles.borderWidth),
+        ),
+        child: Row(
+          children: <Widget>[
+            SizedBox(width: 35.0, child: BaseStyles.iconPrefix(materialIcon)),
+            Expanded(
+              child: Center(
+                child: DropdownButton<String>(
+                  items: buildMaterialItems(items),
+                  value: value,
+                  hint: Text(hintText, style: TextStyles.suggestion),
+                  style: TextStyles.body,
+                  underline: const SizedBox.shrink(),
+                  iconEnabledColor: AppColors.straw,
+                  onChanged: onChanged,
+                ),
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
   }
 
   List<DropdownMenuItem<String>> buildMaterialItems(List<String> items) {
-    return (items != null) ? items
+    return items
         .map((item) => DropdownMenuItem<String>(
-              child: Text(
-                item,
-                textAlign: TextAlign.center,
-              ),
               value: item,
+              child: Text(item, textAlign: TextAlign.center),
             ))
-        .toList() : [];
+        .toList();
   }
 
   List<Widget> buildCupertinoItems(List<String> items) {
@@ -111,20 +118,22 @@ class AppDropdownButton extends StatelessWidget {
         .toList();
   }
 
-  _selectIOS(BuildContext context, List<String> items, String value) {
+  Widget _selectIOS(BuildContext context) {
+    final initialIndex =
+        (value != null) ? items.indexWhere((item) => item == value) : 0;
+
     return GestureDetector(
-      onTap: () {
-        Navigator.of(context).pop();
-      },
+      onTap: () => Navigator.of(context).pop(),
       child: Container(
         color: Colors.white,
         height: 200.0,
         child: CupertinoPicker(
-          scrollController: FixedExtentScrollController(initialItem: items.indexWhere((item) => item == value)),
+          scrollController: FixedExtentScrollController(
+              initialItem: initialIndex < 0 ? 0 : initialIndex),
           itemExtent: 45.0,
           children: buildCupertinoItems(items),
           diameterRatio: 1.0,
-          onSelectedItemChanged: (int index)  => onChanged(items[index]),
+          onSelectedItemChanged: (index) => onChanged?.call(items[index]),
         ),
       ),
     );
