@@ -12,41 +12,49 @@ class AppListTile extends StatelessWidget {
   final bool acceptingOrders;
   final String marketId;
 
-  AppListTile({
-    @required this.month,
-    @required this.date,
-    @required this.title,
-    @required this.location,
-    @required this.marketId,
-    this.acceptingOrders = false
+  const AppListTile({
+    super.key,
+    required this.month,
+    required this.date,
+    required this.title,
+    required this.location,
+    required this.marketId,
+    this.acceptingOrders = false,
   });
 
   @override
   Widget build(BuildContext context) {
-      return Column(
-        children: <Widget>[
-          ListTile( 
-            leading: CircleAvatar(  
-              radius: 25.0,
-              backgroundColor: AppColors.lightblue,
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: <Widget>[
-                  Text(month,style: TextStyle(color:Colors.white, fontSize: 12.0),),
-                  Text(date, style: TextStyles.buttonTextLight)
-                ],),
+    return Column(
+      children: <Widget>[
+        ListTile(
+          leading: CircleAvatar(
+            radius: 25.0,
+            backgroundColor: AppColors.lightblue,
+            child: Column(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: <Widget>[
+                Text(month,
+                    style:
+                        const TextStyle(color: Colors.white, fontSize: 12.0)),
+                Text(date, style: TextStyles.buttonTextLight),
+              ],
             ),
-            title: Text(title,style:TextStyles.subtitle),
-            subtitle: Text(location),
-            trailing:(acceptingOrders) ? Icon(FontAwesomeIcons.shoppingBasket,color: AppColors.darkblue,): Text(''),
-            onTap: (acceptingOrders) ? () => Navigator.of(context).pushNamed('/customer/$marketId'): null,
           ),
-          Padding(
-            padding: EdgeInsets.symmetric(horizontal: BaseStyles.listFieldHorizontal),
-            child: Divider(color:AppColors.lightgray),
-          )
-        ],
-      );
-
+          title: Text(title, style: TextStyles.subtitle),
+          subtitle: Text(location),
+          trailing: acceptingOrders
+              ? Icon(FontAwesomeIcons.basketShopping, color: AppColors.darkblue)
+              : const SizedBox.shrink(),
+          onTap: acceptingOrders
+              ? () => Navigator.of(context).pushNamed('/customer/$marketId')
+              : null,
+        ),
+        Padding(
+          padding:
+              EdgeInsets.symmetric(horizontal: BaseStyles.listFieldHorizontal),
+          child: Divider(color: AppColors.lightgray),
+        )
+      ],
+    );
   }
 }
