@@ -7,41 +7,27 @@ import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 class AppSocialButton extends StatelessWidget {
   final SocialType socialType;
 
-  AppSocialButton({@required this.socialType});
+  const AppSocialButton({super.key, required this.socialType});
 
   @override
   Widget build(BuildContext context) {
-    Color buttonColor;
-    Color iconColor;
-    IconData icon;
-
-    switch (socialType) {
-      case SocialType.Facebook:
-        iconColor = Colors.white;
-        buttonColor = AppColors.facebook;
-        icon = FontAwesomeIcons.facebookF;
-        break;
-      case SocialType.Google:
-        iconColor = Colors.white;
-        buttonColor = AppColors.google;
-        icon = FontAwesomeIcons.google;
-        break;
-      default:
-        iconColor = Colors.white;
-        buttonColor = AppColors.facebook;
-        icon = FontAwesomeIcons.facebookF;
-        break;
-    }
+    final buttonColor = socialType == SocialType.Facebook
+        ? AppColors.facebook
+        : AppColors.google;
+    final icon = socialType == SocialType.Facebook
+        ? FontAwesomeIcons.facebookF
+        : FontAwesomeIcons.google;
 
     return Container(
-        height: ButtonStyles.buttonHeight,
-        width: ButtonStyles.buttonHeight,
-        decoration: BoxDecoration(
-            color: buttonColor,
-            borderRadius: BorderRadius.circular(BaseStyles.borderRadius),
-            boxShadow: BaseStyles.boxShadow),
-        child: Icon(icon, color: iconColor),
-      );
+      height: ButtonStyles.buttonHeight,
+      width: ButtonStyles.buttonHeight,
+      decoration: BoxDecoration(
+        color: buttonColor,
+        borderRadius: BorderRadius.circular(BaseStyles.borderRadius),
+        boxShadow: BaseStyles.boxShadow,
+      ),
+      child: Icon(icon, color: Colors.white),
+    );
   }
 }
 
