@@ -3,22 +3,23 @@ import 'package:farmers_market/src/styles/textfields.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
-class AppTextField extends StatefulWidget{
+class AppTextField extends StatefulWidget {
   final bool isIOS;
   final String hintText;
   final IconData materialIcon;
   final IconData cupertinoIcon;
   final TextInputType textInputType;
   final bool obscureText;
-  final void Function(String) onChanged;
-  final String errorText;
-  final String initialText;
+  final ValueChanged<String>? onChanged;
+  final String? errorText;
+  final String? initialText;
 
-  AppTextField({ 
-    @required this.isIOS,
-    @required this.hintText,
-    @required this.cupertinoIcon,
-    @required this.materialIcon,
+  const AppTextField({
+    super.key,
+    required this.isIOS,
+    required this.hintText,
+    required this.cupertinoIcon,
+    required this.materialIcon,
     this.textInputType = TextInputType.text,
     this.obscureText = false,
     this.onChanged,
@@ -27,34 +28,27 @@ class AppTextField extends StatefulWidget{
   });
 
   @override
-  _AppTextFieldState createState() => _AppTextFieldState();
+  State<AppTextField> createState() => _AppTextFieldState();
 }
 
 class _AppTextFieldState extends State<AppTextField> {
-  FocusNode _node;
-  bool displayCupertinoErrorBorder;
-  TextEditingController _controller;
-
+  late FocusNode _node;
+  bool displayCupertinoErrorBorder = false;
+  late TextEditingController _controller;
 
   @override
   void initState() {
     _node = FocusNode();
-    _controller = TextEditingController();
-    if (widget.initialText != null) _controller.text = widget.initialText;
+    _controller = TextEditingController(text: widget.initialText ?? '');
     _node.addListener(_handleFocusChange);
-    displayCupertinoErrorBorder = false;
     super.initState();
   }
 
-  void _handleFocusChange(){
-    if (_node.hasFocus==false  && widget.errorText != null){
-      displayCupertinoErrorBorder = true;
-    } else {
-      displayCupertinoErrorBorder = false;
-    }
-
-    widget.onChanged(_controller.text);
-
+  void _handleFocusChange() {
+    setState(() {
+      displayCupertinoErrorBorder = !_node.hasFocus && widget.errorText != null;
+    });
+    widget.onChanged?.call(_controller.text);
   }
 
   @override
@@ -67,47 +61,65 @@ class _AppTextFieldState extends State<AppTextField> {
 
   @override
   Widget build(BuildContext context) {
-      if (widget.isIOS){
+    if (widget.isIOS) {
       return Padding(
-        padding: EdgeInsets.symmetric(horizontal: TextFieldStyles.textBoxHorizontal, vertical: TextFieldStyles.textBoxVertical),
+        padding: EdgeInsets.symmetric(
+          horizontal: TextFieldStyles.textBoxHorizontal,
+          vertical: TextFieldStyles.textBoxVertical,
+        ),
         child: Column(
           children: <Widget>[
-            CupertinoTextField(  
+            CupertinoTextField(
               keyboardType: widget.textInputType,
-              padding: EdgeInsets.all(12.0),
+              padding: const EdgeInsets.all(12.0),
               placeholder: widget.hintText,
               placeholderStyle: TextFieldStyles.placeholder,
               style: TextFieldStyles.text,
               textAlign: TextFieldStyles.textAlign,
               cursorColor: TextFieldStyles.cursorColor,
-              decoration: (displayCupertinoErrorBorder) ? TextFieldStyles.cupertinoErrorDecoration : TextFieldStyles.cupertinoDecoration,
+              decoration: displayCupertinoErrorBorder
+                  ? TextFieldStyles.cupertinoErrorDecoration
+                  : TextFieldStyles.cupertinoDecoration,
               prefix: TextFieldStyles.iconPrefix(widget.cupertinoIcon),
               obscureText: widget.obscureText,
               onChanged: widget.onChanged,
               focusNode: _node,
               controller: _controller,
             ),
-            (widget.errorText !=null) ? Padding(
-              padding: const EdgeInsets.only(top: 5.0,left:10.0),
-              child: Row(children: <Widget>[Text(widget.errorText,style: TextStyles.error,)],),
-            ) : Container()
+            widget.errorText != null
+                ? Padding(
+                    padding: const EdgeInsets.only(top: 5.0, left: 10.0),
+                    child: Row(
+                      children: <Widget>[
+                        Text(widget.errorText!, style: TextStyles.error),
+                      ],
+                    ),
+                  )
+                : const SizedBox.shrink(),
           ],
         ),
       );
-    } else {
-      return Padding(
-        padding: EdgeInsets.symmetric(horizontal: TextFieldStyles.textBoxHorizontal, vertical: TextFieldStyles.textBoxVertical),
-        child: TextField(  
-          keyboardType: widget.textInputType,
-          cursorColor: TextFieldStyles.cursorColor,
-          style:TextFieldStyles.text,
-          textAlign: TextFieldStyles.textAlign,
-          decoration: TextFieldStyles.materialDecoration(widget.hintText, widget.materialIcon,widget.errorText),
-          obscureText: widget.obscureText,
-          controller: _controller,
-          onChanged: widget.onChanged,
-        ),
-      );
     }
+
+    return Padding(
+      padding: EdgeInsets.symmetric(
+        horizontal: TextFieldStyles.textBoxHorizontal,
+        vertical: TextFieldStyles.textBoxVertical,
+      ),
+      child: TextField(
+        keyboardType: widget.textInputType,
+        cursorColor: TextFieldStyles.cursorColor,
+        style: TextFieldStyles.text,
+        textAlign: TextFieldStyles.textAlign,
+        decoration: TextFieldStyles.materialDecoration(
+          widget.hintText,
+          widget.materialIcon,
+          widget.errorText,
+        ),
+        obscureText: widget.obscureText,
+        controller: _controller,
+        onChanged: widget.onChanged,
+      ),
+    );
   }
 }
