@@ -2,16 +2,19 @@ class ApplicationUser {
   final String userId;
   final String email;
 
-  ApplicationUser({this.email, this.userId});
+  ApplicationUser({required this.email, required this.userId});
 
-  Map<String,dynamic> toMap(){
+  Map<String, dynamic> toMap() {
     return {
       'userId': userId,
-      'email': email
+      'email': email,
     };
   }
 
-  ApplicationUser.fromFirestore(Map<String,dynamic> firestore)
-    : userId = firestore['userId'],
-      email = firestore['email'];
+  factory ApplicationUser.fromFirestore(Map<String, dynamic> firestore) {
+    return ApplicationUser(
+      userId: (firestore['userId'] ?? '').toString(),
+      email: (firestore['email'] ?? '').toString(),
+    );
+  }
 }
