@@ -43,21 +43,20 @@ abstract class TextStyles {
         textStyle: TextStyle(color: AppColors.darkgray, fontSize: 16.0));
   }
 
-    static TextStyle get bodyLightBlue {
+  static TextStyle get bodyLightBlue {
     return GoogleFonts.roboto(
         textStyle: TextStyle(color: AppColors.lightblue, fontSize: 16.0));
   }
 
-    static TextStyle get bodyRed {
+  static TextStyle get bodyRed {
     return GoogleFonts.roboto(
         textStyle: TextStyle(color: AppColors.red, fontSize: 16.0));
   }
 
-   static TextStyle get picker {
+  static TextStyle get picker {
     return GoogleFonts.roboto(
         textStyle: TextStyle(color: AppColors.darkgray, fontSize: 35.0));
   }
-
 
   static TextStyle get link {
     return GoogleFonts.roboto(
