@@ -1,13 +1,11 @@
-import 'package:cupertino_toolbar/cupertino_toolbar.dart';
-import 'package:farmers_market/src/app.dart';
+import 'dart:io';
+
 import 'package:farmers_market/src/blocs/auth_bloc.dart';
 import 'package:farmers_market/src/blocs/product_bloc.dart';
 import 'package:farmers_market/src/models/product.dart';
 import 'package:farmers_market/src/styles/colors.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'dart:io';
-
 import 'package:provider/provider.dart';
 
 import 'card.dart';
@@ -15,101 +13,71 @@ import 'card.dart';
 class Products extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    var productBloc = Provider.of<ProductBloc>(context);
-    var authBloc = Provider.of<AuthBloc>(context);
+    final productBloc = Provider.of<ProductBloc>(context);
+    final authBloc = Provider.of<AuthBloc>(context);
 
-  return pageBody(productBloc, context, authBloc.userId);
-
+    return Scaffold(
+      body: pageBody(productBloc, context, authBloc.userId),
+      floatingActionButton: FloatingActionButton(
+        backgroundColor: AppColors.straw,
+        child: Icon(Platform.isIOS ? CupertinoIcons.add : Icons.add),
+        onPressed: () => Navigator.of(context).pushNamed('/editproduct'),
+      ),
+    );
+  }
 
   Widget pageBody(
-      ProductBloc productBloc, BuildContext context, String vendorId) {
-    return StreamBuilder<List<Product>>(
-        stream: productBloc.productByVendorId(vendorId),
-        builder: (context, snapshot) {
-          if (!snapshot.hasData)
-            return (Platform.isIOS)
-                ? CupertinoActivityIndicator()
-                : CircularProgressIndicator();
-
-          return Column(
-            children: <Widget>[
-              Expanded(
-                child: ListView.builder(
-                    itemCount: snapshot.data.length,
-                    itemBuilder: (context, index) {
-                      var product = snapshot.data[index];
-                      return GestureDetector(
-                        child: AppCard(
-                          availableUnits: product.availableUnits,
-                          price: product.unitPrice,
-                          productName: product.productName,
-                          unitType: product.unitType,
-                          imageUrl: product.imageUrl,
-                        ),
-                        onTap: () => Navigator.of(context)
-                            .pushNamed('/editproduct/${product.productId}'),
-                      );
-                    }),
-              ),
-              GestureDetector(
-                child: Container(
-                  height: 50.0,
-                  width: double.infinity,
-                  color: AppColors.straw,
-                  child: (Platform.isIOS)
-                      ? Icon(CupertinoIcons.add,
-                          color: Colors.white, size: 35.0)
-                      : Icon(Icons.add, color: Colors.white, size: 35.0),
-                ),
-                onTap: () => Navigator.of(context).pushNamed('/editproduct'),
-              )
-            ],
-          );
-        });
-    if (Platform.isIOS) {
-      return CupertinoPageScaffold(
-        child: CupertinoToolbar(
-          items: <CupertinoToolbarItem>[
-            CupertinoToolbarItem(
-                icon: CupertinoIcons.add_circled,
-                onPressed: () =>
-                    Navigator.of(context).pushNamed('/editproduct'))
-          ],
-          body: pageBody(productBloc,context,authBloc.userId),
-        ),
-      );
-    } else {
-      return Scaffold(
-        body: pageBody(productBloc,context,authBloc.userId),
-        floatingActionButton: FloatingActionButton(
-          backgroundColor: AppColors.straw,
-          child: Icon(Icons.add),
-          onPressed: () => Navigator.of(context).pushNamed('/editproduct'),
+      ProductBloc productBloc, BuildContext context, String? vendorId) {
+    if (vendorId == null || vendorId.isEmpty) {
+      return Center(
+        child: Text(
+          'Vendor belum ditemukan.',
+          style: TextStyle(color: AppColors.darkgray),
         ),
       );
     }
-  }
 
-  Widget pageBody(ProductBloc productBloc, BuildContext context, String vendorId) {
     return StreamBuilder<List<Product>>(
       stream: productBloc.productByVendorId(vendorId),
       builder: (context, snapshot) {
-        if (!snapshot.hasData) return (Platform.isIOS)
-        ? CupertinoActivityIndicator()
-        : CircularProgressIndicator();
+        if (!snapshot.hasData) {
+          return Center(
+            child: Platform.isIOS
+                ? const CupertinoActivityIndicator()
+                : const CircularProgressIndicator(),
+          );
+        }
+
+        final products = snapshot.data!;
+        if (products.isEmpty) {
+          return Center(
+            child: Text(
+              'Belum ada produk. Tekan tombol + untuk tambah produk baru.',
+              textAlign: TextAlign.center,
+              style: TextStyle(color: AppColors.darkgray, fontSize: 16.0),
+            ),
+          );
+        }
 
         return ListView.builder(
-          itemCount: snapshot.data.length,
-          itemBuilder: (context, index){
-            var product = snapshot.data[index];
-            return AppCard(
-              availableUnits: product.availableUnits,
-              price: product.unitPrice,
-              productName: product.productName,
-              unitType: product.unitType, 
+          itemCount: products.length,
+          itemBuilder: (context, index) {
+            final product = products[index];
+            return GestureDetector(
+              child: AppCard(
+                availableUnits: product.availableUnits,
+                price: product.unitPrice,
+                productName: product.productName,
+                unitType: product.unitType,
+                imageUrl: product.imageUrl,
+                note: product.note,
+              ),
+              onTap: () => Navigator.of(context)
+                  .pushNamed('/editproduct/${product.productId}'),
             );
-          });
-      }
+          },
+        );
+      },
     );
   }
 }
