@@ -1,46 +1,28 @@
-import 'package:farmers_market/src/blocs/auth_bloc.dart';
-import 'package:flutter/cupertino.dart';
-import 'package:flutter/material.dart';
 import 'dart:io';
 
-import 'package:provider/provider.dart';
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
 
 class Profile extends StatelessWidget {
+  const Profile({super.key});
+
   @override
   Widget build(BuildContext context) {
+    final body = Center(
+      child: Platform.isIOS
+          ? CupertinoButton(
+              child: const Text('Add'),
+              onPressed: () => Navigator.of(context).pushNamed('/editvendor'),
+            )
+          : TextButton(
+              child: const Text('Add'),
+              onPressed: () => Navigator.of(context).pushNamed('/editvendor'),
+            ),
+    );
+
     if (Platform.isIOS) {
-      return CupertinoPageScaffold(
-        child: pageBody(context),
-      );
-    } else {
-      return Scaffold(
-        body: pageBody(context),
-      );
+      return CupertinoPageScaffold(child: body);
     }
-  }
-
-  Widget pageBody(BuildContext context) {
-    var authBloc = Provider.of<AuthBloc>(context);
-    // return Center(
-    //     child: (Platform.isIOS)
-    //         ? CupertinoButton(
-    //             child: Text('Logout'),
-    //             onPressed: () => authBloc.logout(),
-    //           )
-    //         : FlatButton(
-    //             child: Text('Logout'),
-    //             onPressed: () => authBloc.logout(),
-    //           ));
-
-              return Center(
-        child: (Platform.isIOS)
-            ? CupertinoButton(
-                child: Text('Add'),
-                onPressed: () => Navigator.of(context).pushNamed('/editvendor') ,
-              )
-            : FlatButton(
-                child: Text('Add'),
-                onPressed: () => Navigator.of(context).pushNamed('/editvendor'),
-              ));
+    return Scaffold(body: body);
   }
 }
