@@ -3,8 +3,6 @@ import 'package:flutter/material.dart';
 class ShoppingBag extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return Container(
-      
-    );
+    return Container();
   }
 }
