@@ -8,35 +8,33 @@ abstract class VendorScaffold {
   static CupertinoTabScaffold get cupertinoTabScaffold {
     return CupertinoTabScaffold(
       tabBar: _cupertinoTabBar,
-      tabBuilder: (context, index) {
-        return _pageSelection(index);
-      },
+      tabBuilder: (context, index) => _pageSelection(index),
     );
   }
 
-  static get _cupertinoTabBar {
+  static CupertinoTabBar get _cupertinoTabBar {
     return CupertinoTabBar(
       backgroundColor: AppColors.darkblue,
-      items: <BottomNavigationBarItem>[
+      items: const <BottomNavigationBarItem>[
         BottomNavigationBarItem(
-            icon: Icon(CupertinoIcons.create), title: Text('Products')),
+          icon: Icon(CupertinoIcons.create),
+          label: 'Products',
+        ),
         BottomNavigationBarItem(
-            icon: Icon(CupertinoIcons.shopping_cart), title: Text('Orders')),
+          icon: Icon(CupertinoIcons.shopping_cart),
+          label: 'Orders',
+        ),
         BottomNavigationBarItem(
-            icon: Icon(CupertinoIcons.person), title: Text('Profile')),
+          icon: Icon(CupertinoIcons.person),
+          label: 'Profile',
+        ),
       ],
     );
   }
 
   static Widget _pageSelection(int index) {
-    if (index == 0) {
-      return Products();
-    }
-
-    if (index == 1) {
-      return Orders();
-    }
-
-    return Profile();
+    if (index == 0) return Products();
+    if (index == 1) return Orders();
+    return const Profile();
   }
 }
